Extract classes from src/providers/preview/mathpreview.ts to src/providers/preview/mathpreviewlib/
