+++ src/providers/preview/mathpreviewlib/cursorrenderer.ts
@@ -1,0 +1,40 @@
+import * as vscode from 'vscode'
+
+
+export class CursorRenderer {
+
+    // Test whether cursor is in tex command strings
+    // like \begin{...} \end{...} \xxxx{ \[ \] \( \) or \\
+    private isCursorInTeXCommand(document: vscode.TextDocument): boolean {
+        const editor = vscode.window.activeTextEditor
+        if (!editor) {
+            return false
+        }
+        const cursor = editor.selection.active
+        const r = document.getWordRangeAtPosition(cursor, /\\(?:begin|end|label)\{.*?\}|\\[a-zA-Z]+\{?|\\[()[\]]|\\\\/)
+        if (r && r.start.isBefore(cursor) && r.end.isAfter(cursor) ) {
+            return true
+        }
+        return false
+    }
+
+    renderCursor(document: vscode.TextDocument, range: vscode.Range, thisColor: string): string {
+        const editor = vscode.window.activeTextEditor
+        const configuration = vscode.workspace.getConfiguration('latex-workshop')
+        const conf = configuration.get('hover.preview.cursor.enabled') as boolean
+        if (editor && conf && !this.isCursorInTeXCommand(document)) {
+            const cursor = editor.selection.active
+            const symbol = configuration.get('hover.preview.cursor.symbol') as string
+            const color = configuration.get('hover.preview.cursor.color') as string
+            let sym = `{\\color{${thisColor}}${symbol}}`
+            if (color !== 'auto') {
+                sym = `{\\color{${color}}${symbol}}`
+            }
+            if (range.contains(cursor) && !range.start.isEqual(cursor) && !range.end.isEqual(cursor)) {
+                return document.getText( new vscode.Range(range.start, cursor) ) + sym + document.getText( new vscode.Range(cursor, range.end))
+            }
+        }
+        return document.getText(range)
+    }
+
+}
