+++ src/providers/preview/mathpreviewlib/textdocumentlike.ts
